fixed bugs that were found when building blog

+++ server/timer.h
@@ -8,7 +8,7 @@ typedef void (*timer_handler_t)(void *);
 #define FREE_LATER(ptr) if (ptr) set_timeout(0, free, ptr)
 
 void explode_timers(void);
-int get_min_timeout(void);
+i64 get_min_timeout(void);
 void show_timers(void);
 void *set_interval(u32 seconds, timer_handler_t handler, void *arg);
 void *set_timeout(u32 seconds, timer_handler_t handler, void *arg);
